requirements.txt
-e .
aws-cdk.core==1.99.0
aws-cdk.aws_iam==1.99.0
aws-cdk.aws_s3==1.99.0
aws_cdk.custom_resources==1.99.0
aws-cdk.aws-kinesis==1.99.0
aws-cdk.aws-kinesisfirehose==1.99.0
aws-cdk.aws_lambda==1.99.0

